tools automatic update with Git, Shell bat file
In this repo, I implemented a mechanism to automatically update the tools (or any other files) with Git.  The Git commands are run under Cygwin environment. What's more, a simple b shell script is written to run these Git commands. Futhermode, this shell script is run automatically by setting up a task in Taskscheduler in Windows